requirements.txt
numpy==1.24.3
matplotlib==3.7.1
PyYAML==6.0
dash==2.9.3
plotly==5.14.1
gunicorn==20.1.0